+++ changelog.md
@@ -4,6 +4,14 @@
 
 Initial upload
 
-- December ?, 2018
+- June 18, 2018 - 1.0
 
-Added Lucene
+Added Lucene (searching is much faster, but initial indexing is slower)
+Added search by author name
+Finished prequel/sequel finder
+
+- June 19, 2018 - 1.01
+
+Fixed write to file
+Can search by content rating
+Small improvements